Motion Detector working! Al though a bit sensitive. Will be working on making it less sensitive, finding the right parameters for it and such.
For motion_detector.py, commented out the elif and else in the if else block for assigning frames. Changed functionality of script to test current frame against first background frame. This helps to detect minute changes. Before, comparing the 2 most recent frames, it would not pick up slow motions.
Added a Snippet folder to contain all of the small scripts used for small experimental use.
Added the BC folder that contains the virtualenv for the project, as well as BC.txt which is just a txt file that has the command needed to activate the virtenv.